Create font management and add score
- Add font management and text textures
- Add score and countdown
- Centralized configuration management
- Textures color modulation
- NumberPrinter added
- Countdown and ScoreDisplay added
- Sprite: pos/scale management isolated, drawing is delegated to ISpriteDrawer
- IAnimated added, normalisation of all animations updates
- Number printer looks like other resources and is isolated from its sprite drawer
- Numbers shadows added to game configuration

+++ src/main.cpp
@@ -51,5 +51,5 @@ int main(int argc, char** argv)
 
 	//Run game
 	game::MineDigger game;
-	return static_cast<int>(app::run(game));
+	return app::run(game);
 }
